Change review system for books

+++ books/views.py
@@ -1,18 +1,30 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
-from django.contrib.auth.decorators import login_required
-from django.db.models import Q
-from django.db.models.functions import Lower
-
+from django.db.models import Q, Case, When, F, DecimalField, Avg, Value
 from .models import Book, Category
+from django.db.models.functions import Lower
+from django.http import JsonResponse
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.models import User
+from profiles.models import UserProfile
+from reviews.models import Review
+from checkout.models import Order
 from .forms import BookForm
+from django.http import HttpResponseBadRequest
+from datetime import date
 
-# Create your views here.
 
 def all_books(request):
-    """ A view to show all books, including sorting and search queries """
+    """
+    All books
+    """
+    books = Book.objects.annotate(
+        average_rating=Avg('reviews__rating')
+    )
 
-    books = Book.objects.all()
+    for book in books:
+        book.average_rating = book.calculate_average_rating()
+
     query = None
     categories = None
     sort = None
@@ -24,15 +36,50 @@ def all_books(request):
             sort = sortkey
             if sortkey == 'name':
                 sortkey = 'lower_name'
-                books = books.annotate(lower_name=Lower('name'))
+                books = books.annotate(
+                    lower_name=Lower('name')
+                    )
             if sortkey == 'category':
                 sortkey = 'category__name'
             if 'direction' in request.GET:
                 direction = request.GET['direction']
+            if sortkey == 'price':
+                if direction == 'asc':
+                    books = books.annotate(
+                        sorted_price=Case(
+                            default='price',
+                            output_field=DecimalField()
+                        )
+                    ).order_by(F('sorted_price').asc(nulls_last=True))
+                elif direction == 'desc':
+                    books = books.annotate(
+                        sorted_price=Case(
+                            default='price',
+                            output_field=DecimalField()
+                        )
+                    ).order_by(F('sorted_price').desc(nulls_last=True))
+            elif sortkey == 'rating':
+                if direction == 'asc':
+                    books = books.annotate(
+                        rating_is_null=Case(
+                            When(average_rating__isnull=True, then=Value(1)),
+                            default=Value(0),
+                            output_field=DecimalField()
+                        )
+                    ).order_by('rating_is_null', 'average_rating')
+                elif direction == 'desc':
+                    books = books.annotate(
+                        rating_is_null=Case(
+                            When(average_rating__isnull=True, then=Value(1)),
+                            default=Value(0),
+                            output_field=DecimalField()
+                        )
+                    ).order_by('rating_is_null', '-average_rating')
+            else:
                 if direction == 'desc':
                     sortkey = f'-{sortkey}'
-            books = books.order_by(sortkey)
-            
+                books = books.order_by(sortkey)
+
         if 'category' in request.GET:
             categories = request.GET['category'].split(',')
             books = books.filter(category__name__in=categories)
@@ -41,10 +88,14 @@ def all_books(request):
         if 'q' in request.GET:
             query = request.GET['q']
             if not query:
-                messages.error(request, "You didn't enter any search criteria!")
+                messages.error(
+                    request,
+                    "You didn't enter any search criteria!"
+                )
                 return redirect(reverse('books'))
-            #Filter books 
-            queries = Q(title__icontains=query) | Q(description__icontains=query)
+
+            queries = \
+                Q(title__icontains=query) | Q(description__icontains=query)
             books = books.filter(queries)
 
     current_sorting = f'{sort}_{direction}'
@@ -60,22 +111,82 @@ def all_books(request):
 
 
 def book_detail(request, book_id):
-    """ A view to show individual book details """
+    """
+    Book details
 
+    """
     book = get_object_or_404(Book, pk=book_id)
+    reviews = book.reviews.all()
+    average_rating = book.calculate_average_rating()
+    user = request.user
+    has_ordered_book = False
+    has_reviewed = False
+
+    if user.is_authenticated:
+        has_ordered_book = Order.objects.filter(
+            user_profile=user.userprofile, lineitems__book=book
+        ).exists()
+        has_reviewed = Review.objects.filter(
+            reviewer=user, book=book
+        ).exists()
 
     context = {
         'book': book,
+        'reviews': reviews,
+        'average_rating': average_rating,
+        'has_ordered_book': has_ordered_book,
+        'has_reviewed': has_reviewed,
     }
 
     return render(request, 'books/book_detail.html', context)
 
 
 @login_required
+def wishlist(request, sku, *args, **kwargs):
+    """
+    book whislist
+    """
+    book_wish = get_object_or_404(Book, sku=sku)
+    user = request.user
+    user_profile = user.userprofile
+
+    liked = False
+
+    if book_wish.wishlist.filter(id=request.user.id).exists():
+        book_wish.wishlist.remove(request.user)
+        user_profile.wishlist.remove(book_wish)
+        messages.success(
+            request,
+            f'Successfully removed {book_wish} to Wishlist!'
+        )
+        liked = False
+    else:
+        book_wish.wishlist.add(request.user)
+        user_profile.wishlist.add(book_wish)
+        messages.success(
+            request, f'Successfully added {book_wish} to Wishlist!'
+        )
+        liked = True
+
+    return JsonResponse(
+        {
+            'wishlist_count': book_wish.wishlist.count(),
+            'liked': liked
+        }
+    )
+
+
+@login_required
 def add_book(request):
-    """ Add a book to the store """
+    """
+    Add a book to the store.
+ 
+    """
     if not request.user.is_superuser:
-        messages.error(request, 'Sorry, only store owners can do that.')
+        messages.error(
+            request,
+            'Sorry, only store owners can do that.'
+        )
         return redirect(reverse('home'))
 
     if request.method == 'POST':
@@ -85,10 +196,12 @@ def add_book(request):
             messages.success(request, 'Successfully added book!')
             return redirect(reverse('book_detail', args=[book.id]))
         else:
-            messages.error(request, 'Failed to add book. Please ensure the form is valid.')
+            messages.error(
+                request,
+                'Failed to add book. Please ensure the form is valid.'
+            )
     else:
         form = BookForm()
-        
     template = 'books/add_book.html'
     context = {
         'form': form,
@@ -99,9 +212,14 @@ def add_book(request):
 
 @login_required
 def edit_book(request, book_id):
-    """ Edit a book in the store """
+    """
+    Edit a book
+    """
     if not request.user.is_superuser:
-        messages.error(request, 'Sorry, only store owners can do that.')
+        messages.error(
+            request,
+            'Sorry, only admins can do that.'
+        )
         return redirect(reverse('home'))
 
     book = get_object_or_404(Book, pk=book_id)
@@ -112,7 +230,10 @@ def edit_book(request, book_id):
             messages.success(request, 'Successfully updated book!')
             return redirect(reverse('book_detail', args=[book.id]))
         else:
-            messages.error(request, 'Failed to update book. Please ensure the form is valid.')
+            messages.error(
+                request,
+                'Failed to update book. Please ensure the form is valid.'
+            )
     else:
         form = BookForm(instance=book)
         messages.info(request, f'You are editing {book.title}')
@@ -128,12 +249,50 @@ def edit_book(request, book_id):
 
 @login_required
 def delete_book(request, book_id):
-    """ Delete a book from the store """
+    """
+    Delete a book
+    """
     if not request.user.is_superuser:
-        messages.error(request, 'Sorry, only store owners can do that.')
+        messages.error(
+            request,
+            'Sorry, only adminss can do that.'
+        )
         return redirect(reverse('home'))
 
-    book = get_object_or_404(Book, pk=book_id)
-    book.delete()
-    messages.success(request, 'Book deleted!')
-    return redirect(reverse('books'))
+    if request.method == 'POST':
+        book_id = request.POST.get('book_id')
+        if book_id:
+            book = get_object_or_404(Book, pk=book_id)
+            book.delete()
+            messages.success(request, 'Book deleted successfully.')
+            return redirect('books')
+        else:
+            return HttpResponseBadRequest('Invalid request')
+    else:
+        return HttpResponseBadRequest('Invalid request')
+
+
+
+
+
+@login_required
+def admin_books(request):
+    """
+    View all books for administrators.
+    """
+    if not request.user.is_superuser:
+        messages.error(
+            request,
+            'Sorry, only administrators can access this page.'
+        )
+        return redirect(reverse('home'))
+
+    books = Book.objects.all()
+
+    context = {
+        'books': books
+    }
+    return render(request, 'books/admin_books.html', context)
+
+
+
